mergesort finally compiles but segfaults when run, others run except maplist doesn't access v1 through the env

+++ outputs/sumListCall_output.c
@@ -32,7 +32,7 @@ Closure* v5(int v2) {
 
 int v0(Node* v1) {
   Node* v6 = v1;
-  return (isEmpty(v6)) ? (0) : ((int)(intptr_t)apply(v5(*(int*)head(v6)), tail(v6)));
+  return (isEmpty(v6)) ? (0) : (apply(v5(*(int*)head(v6)), tail(v6)));
 }
 
 // main
